+++ node/src/main/kotlin/net/corda/node/serialization/kryo/KryoServerSerializationScheme.kt
@@ -1,9 +1,8 @@
-package net.corda.node.serialization
+package net.corda.node.serialization.kryo
 
 import com.esotericsoftware.kryo.pool.KryoPool
 import net.corda.core.serialization.SerializationContext
 import net.corda.nodeapi.internal.serialization.CordaSerializationMagic
-import net.corda.node.services.messaging.RpcServerObservableSerializer
 import net.corda.nodeapi.internal.serialization.kryo.AbstractKryoSerializationScheme
 import net.corda.nodeapi.internal.serialization.kryo.DefaultKryoCustomizer
 import net.corda.nodeapi.internal.serialization.kryo.kryoMagic
